CTable indexing code move into its own module
